// frontend/src/api-generated/models/phieuThanhLyTaiSanResponse.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */
import type { ChiTietThanhLyGeneralResponse } from './chiTietThanhLyGeneralResponse';

export interface PhieuThanhLyTaiSanResponse {
  id?: number;
  idDonVi?: number;
  maPhieuThanhLy?: string;
  tenNguoiLap?: string;
  tenNguoiPheDuyet?: string;
  hinhThucThanhLy?: string;
  tongTienThuHoi?: number;
  thoiGianThanhLy?: string;
  trangThaiLucGiao?: string;
  lyDoThanhLy?: string;
  trangThai?: string;
  thoiGianTao?: string;
  thoiGianCapNhat?: string;
  /** Danh sách gộp chung phẳng toàn bộ tài sản thanh lý */
  chiTietTaiSan?: ChiTietThanhLyGeneralResponse[];
}