// frontend/src/api-generated/models/phieuThanhLyTaiSanRequest.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */
import type { ChiTietThanhLyLinhKienRequest } from './chiTietThanhLyLinhKienRequest';
import type { ChiTietThanhLyPhanCungRequest } from './chiTietThanhLyPhanCungRequest';
import type { ChiTietThanhLyPhanMemRequest } from './chiTietThanhLyPhanMemRequest';

export interface PhieuThanhLyTaiSanRequest {
  hinhThucThanhLy?: string;
  tongTienThuHoi?: number;
  trangThaiLucGiao?: string;
  lyDoThanhLy?: string;
  danhSachPhanCung?: ChiTietThanhLyPhanCungRequest[];
  danhSachPhanMem?: ChiTietThanhLyPhanMemRequest[];
  danhSachLinhKien?: ChiTietThanhLyLinhKienRequest[];
}